Refactor navbar and add page components
Moved navbar to components directory and updated imports across pages. Added basic page components for Calculation, Calorie Tracker, My Log, and Routine Recommendation, each rendering the navbar and a header. Updated navbar links to route to the correct pages.
