+++ easyfitness/src/app/calculation/page.tsx
@@ -1,0 +1,11 @@
+import Nav from '../components/navbar';
+
+export default function CalculationPage() {
+    return (
+        <div>
+        <Nav />
+        <h1 className="text-3xl font-bold text-center mt-8">Calculation</h1>
+        <p className="text-center mt-4">This is the Calculation page.</p>
+        </div>
+    );
+    }
