+++ easyfitness/src/app/components/navbar.tsx
@@ -13,22 +13,22 @@ export default function Nav() {
       {/* Center: Navigation Headers */}
       <NavbarContent className="hidden sm:flex gap-6" justify="center">
         <NavbarItem>
-          <Link color="foreground" href="#" className="text-xl px-4 py-2">
+          <Link as={Link} href="/mylog" color="foreground" className="text-xl px-4 py-2">
             My Log
           </Link>
         </NavbarItem>
         <NavbarItem isActive>
-          <Link aria-current="page" href="#" className="text-xl px-4 py-2">
+          <Link aria-current="page" href="/routine" className="text-xl px-4 py-2">
             Routine Recommendation
           </Link>
         </NavbarItem>
         <NavbarItem>
-          <Link color="foreground" href="#" className="text-xl px-4 py-2">
+          <Link color="foreground" href="/calorietrack" className="text-xl px-4 py-2">
             Calorie Tracker
           </Link>
         </NavbarItem>
         <NavbarItem>
-          <Link color="foreground" href="#" className="text-xl px-4 py-2">
+          <Link color="foreground" href="/calculation" className="text-xl px-4 py-2">
             Calculation
           </Link>
         </NavbarItem>
