+++ easyfitness/src/app/mylog/page.tsx
@@ -1,0 +1,11 @@
+import Nav from '../components/navbar';
+
+export default function MyLogPage() {
+    return (
+        <div>
+        <Nav />
+        <h1 className="text-3xl font-bold text-center mt-8">My Log</h1>
+        <p className="text-center mt-4">This is the My Log page.</p>
+        </div>
+    );
+    }
