+++ easyfitness/src/app/routine/page.tsx
@@ -1,0 +1,11 @@
+import Nav from '../components/navbar';
+
+export default function RoutinePage() {
+    return (
+        <div>
+        <Nav />
+        <h1 className="text-3xl font-bold text-center mt-8">Routine Recommendation</h1>
+        <p className="text-center mt-4">This is the Routine Recommendation page.</p>
+        </div>
+    );
+    }
